Move hook to own file
