+++ src/pages/manage/movies/MovieForm.tsx
@@ -1,11 +1,8 @@
-import { yupResolver } from '@hookform/resolvers/yup';
 import { format } from 'date-fns';
 import { useRouter } from 'next/router';
 import { useEffect, useMemo, useState } from 'react';
-import { useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { useQueryClient } from 'react-query';
-import * as yup from 'yup';
 
 import { CalendarType } from '@/constants/CalendarType';
 import { EventTypes } from '@/constants/EventTypes';
@@ -39,6 +36,7 @@ import { parseOfferId } from '@/utils/parseOfferId';
 import { useAddEvent } from './useAddEvent';
 import { useEditField } from './useEditField';
 import { useGetEvent } from './useGetEvent';
+import { useParseStepConfiguration } from './useParseStepConfiguration';
 import { usePublishEvent } from './usePublishEvent';
 import { useToast } from './useToast';
 
@@ -120,48 +118,6 @@ const useFooterStatus = ({ event, form }) => {
   }, [footerStatus]);
 
   return footerStatus;
-};
-
-const useParseStepConfiguration = (
-  configuration,
-  { formConfiguration = {} } = {},
-) => {
-  const schema = yup
-    .object(
-      configuration.reduce(
-        (acc: Object, val: { field: string; validation: () => void }) => {
-          if (!val.field || !val.validation) return acc;
-
-          return {
-            ...acc,
-            [val.field]: val.validation,
-          };
-        },
-        {},
-      ),
-    )
-    .required();
-
-  const resolver = yupResolver(schema);
-
-  const defaultValues = configuration.reduce(
-    (acc: Object, val: { field: string; defaultValue: unknown }) => {
-      if (!val.field || !val.defaultValue) return acc;
-      return {
-        ...acc,
-        [val.field]: val.defaultValue,
-      };
-    },
-    {},
-  );
-
-  const form = useForm<FormData>({
-    resolver,
-    defaultValues,
-    ...formConfiguration,
-  });
-
-  return { form };
 };
 
 const StepForm = ({
